requirements.txt
pip
torch>=0.3
torchvision
imutils==0.4.6
matplotlib
requests>=2.11.1
numpy>=1.14.5
ipython
Pillow>=5.2.0
opencv-python
tox-travis
tqdm
ipywidgets

